requirements.txt
slack-bolt==1.18.1
slack-sdk==3.26.0
kubernetes==28.1.0
python-dateutil==2.8.2
